requirements.txt
fastapi
uvicorn[standard]
jinja2
python-multipart
SQLAlchemy
psycopg2-binary
alembic
python-slugify